apps/server/pyproject.toml
[project]
name = "searchlightai-server"
version = "0.1.0"
description = "video query search appliction"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch==2.8.0+cu128",
    "asyncpg>=0.31.0",
    "fastapi>=0.126.0",
    "faster-whisper>=1.2.1",
    "opencv-python>=4.11.0.86",
    "pydantic-settings>=2.12.0",
    "qdrant-client>=1.16.2",
    "scenedetect>=0.6.7.1",
    "sqlmodel>=0.0.27",
    "structlog>=25.5.0",
    "pillow>=12.0.0",
    "transformers>=4.57.3",
    "sentence-transformers>=5.2.0",
    "python-multipart>=0.0.21",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]
